+++ HISTORY.md
@@ -4,6 +4,8 @@
 
 _????-??-??_
 
+ * Fix command-line duplicate output bug when loading matrices for some bindings
+   (#3838).
 
 ## mlpack 4.5.1
 
@@ -19,9 +21,6 @@ _2024-12-02_
  * Fix compilation if only including `mlpack/methods/kde/kde_model.hpp` (#3800).
 
  * Fix serialization and `MinDistance()` bugs with `HollowBallBound` (#3808).
-
- * Fix command-line duplicate output bug when loading matrices for some bindings
-   (#3838).
 
 ## mlpack 4.5.0