+++ HISTORY.md
@@ -4,6 +4,8 @@
 
 _????-??-??_
 
+ * Fix command-line duplicate output bug when loading matrices for some bindings
+   (#3838).
 
 ## mlpack 4.5.1