+++ HISTORY.md
@@ -15,6 +15,10 @@ _????-??-??_
 
  * Fix serialization and `MinDistance()` bugs with `HollowBallBound` (#3808).
 
+ * Add `LinearRecurrent` layer and revamp `LSTM` layer (#3859).
+
+ * Fix BPTT issues in `RNN` (#3859).
+
 ## mlpack 4.5.0
 
 _2024-09-17_